+++ source/datacontainer.h
@@ -3,9 +3,9 @@
 
 typedef struct dataContainer{
   unsigned char* data;
-  uint64_t bytesize;
+  size_t bytesize;
   
   int (*destroyDataContainer)(struct dataContainer** thisPointer);
 }dataContainer;
 
-dataContainer* newDataContainer(uint64_t bytesize);
+dataContainer* newDataContainer(size_t bytesize);
