+++ source/memorymanager.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <stdint.h>
 
-void *secureAllocate(uint64_t bytesize);
-int secureFree(void *memory, uint64_t bytesize); //NOTE memory is really a void**
+
+void *secureAllocate(size_t bytesize);
+int secureFree(void *memory, size_t bytesize); //NOTE memory is really a void**
